+++ src/enfile.h
@@ -2,6 +2,6 @@
 #define F15_SE2_ENFILE
 /* file helpers (enfile.c) */
 
-void srandInit(int seed);
+void srandInit(int16 seed);
 
 #endif /* F15_SE2_ENFILE */
