+++ src/f15util.h
@@ -7,7 +7,7 @@
 void strcpyFar(const char *src, const uint16 destSegment, const uint16 destOffset, size_t size);
 void memcpyFar(const char *src, const uint16 destSegment, const uint16 destOffset, size_t size);
 void writeWordFar(const uint16 segment, const uint16 offset, const uint16 value);
-int blitFileFar(const char *filename, const uint16 segment, const uint16 offset);
+int16 blitFileFar(const char *filename, const uint16 segment, const uint16 offset);
 const char *sizeString(const size_t size);
 
 #endif // F15UTIL_H
