+++ src/shared/cleanup.c
@@ -9,12 +9,12 @@
 #include <dos.h>
 
 /* extern declarations needed by these functions */
-extern int far gfx_drawString(int16 *pageNum, const char *string);
-extern int far gfx_setFont(uint16 ch, uint16 font);
+extern int16 far gfx_drawString(int16 *pageNum, const char *string);
+extern int16 far gfx_setFont(uint16 ch, uint16 font);
 extern void far misc_clearKeyFlags(void);
 extern uint8 timerHandlerInstalled;
 void restoreTimerIrqHandler(void);
-void intDispatch(int intNum, uint8 *inRegs, uint8 *outRegs);
+void intDispatch(int16 intNum, uint8 *inRegs, uint8 *outRegs);
 
 void cleanup() {
     uint8 regs[0xe];
