+++ src/shared/miscimpl.c
@@ -33,7 +33,7 @@ void mystrcat(char *dst, const char *src) {
     strcat(dst, src);
 }
 
-void intDispatch(int intnum, uint8 *inreg, uint8 *outreg) {
+void intDispatch(int16 intnum, uint8 *inreg, uint8 *outreg) {
     union REGS r;
     /* inreg[0] = AL, inreg[1] = AH based on stinit.c usage */
     r.h.al = inreg[0];
@@ -43,31 +43,31 @@ void intDispatch(int intnum, uint8 *inreg, uint8 *outreg) {
     outreg[1] = r.h.ah;
 }
 
-void doNothing2(const char *msg, int a, int b, int c) {
+void doNothing2(const char *msg, int16 a, int16 b, int16 c) {
 }
 
-int getTimeOfDay(void) {
+int16 getTimeOfDay(void) {
     return 0;
 }
 
-int mystrlen(const char *s) {
+int16 mystrlen(const char *s) {
     return strlen(s);
 }
 
-void nearmemset(void *dst, char val, int count) {
+void nearmemset(void *dst, char val, int16 count) {
     memset(dst, val, count);
 }
 
-int loadOverlay(const char *filename) {
+int16 loadOverlay(const char *filename) {
     return 0;
 }
 
-int doFcbSearch(void) {
+int16 doFcbSearch(void) {
     return -1;
 }
 
 #if !defined(MSDOS)
-uint16 dos_alloc(int size) {
+uint16 dos_alloc(int16 size) {
     return 0;
 }
 #endif
